requirements.txt
mysql-connector-python~=8.0.33
pandas~=1.5.3
matplotlib~=3.5.2

